add country code config
Noticed a maintenance notification to set my country in Home Assistant. Unfortunately I couldn't find any documentation on the YAML to do that exactly except the error when I tried guessing. The error pointed to ISO 3166 Country codes, and a two digit country code seemed to pass validation. So hopefully this is correct